pyproject.toml
[project]
name = "reviewlens-ai"
version = "0.1.0"
description = "Review ingestion and analysis backend for ReviewLens AI."
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.115.0",
    "google-play-scraper",
    "greenlet>=3.1.0",
    "httpx>=0.27.2",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "redis>=5.1.0",
    "rq>=1.16.0",
    "sqlalchemy>=2.0.35",
    "uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "db: requires a reachable PostgreSQL test database",
    "integration: hits live scrape providers",
]

[tool.setuptools.packages.find]
include = ["backend*"]
exclude = ["frontend*", "tests*"]

[tool.ruff]
line-length = 100
target-version = "py311"
